# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DzVêtements Fashion Marketplace API for Algeria
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sellers/{id}:
    get:
      operationId: getSeller
      tags: [sellers]
      summary: Get seller profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sellers/{id}/products:
    get:
      operationId: getSellerProducts
      tags: [sellers]
      summary: Get products by seller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Seller products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        currency:
          type: string
          default: DZD
        images:
          type: array
          items:
            type: string
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        condition:
          type: string
          enum: [new, like_new, good, fair]
        brand:
          type: string
          nullable: true
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerAvatar:
          type: string
          nullable: true
        wilaya:
          type: string
        city:
          type: string
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        viewCount:
          type: integer
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - price
        - currency
        - images
        - category
        - sizes
        - colors
        - condition
        - sellerId
        - sellerName
        - wilaya
        - city
        - reviewCount
        - viewCount
        - isFeatured
        - isNew
        - createdAt
...
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
...
    Seller:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        wilaya:
          type: string
        city:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        productCount:
          type: integer
        isVerified:
          type: boolean
        joinedDate:
          type: string
        responseRate:
          type: number
      required:
        - id
        - name
        - wilaya
        - city
        - rating
        - reviewCount
        - productCount
        - isVerified
        - joinedDate
        - responseRate
    SellerDetail:
      allOf:
        - $ref: "#/components/schemas/Seller"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
...
    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerName:
          type: string
        reviewerAvatar:
          type: string
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - reviewerName
        - rating
        - comment
        - createdAt
...