# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DzVêtements Fashion Marketplace API for Algeria
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCondition } from "./productCondition";

export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  currency: string;
  images: string[];
  category: string;
  subcategory?: string | null;
  sizes: string[];
  colors: string[];
  condition: ProductCondition;
  brand?: string | null;
  sellerId: number;
  sellerName: string;
  sellerAvatar?: string | null;
  wilaya: string;
  city: string;
  rating?: number | null;
  reviewCount: number;
  viewCount: number;
  isFeatured: boolean;
  isNew: boolean;
  createdAt: string;
}